add rspec test

+++ lib/secret-keeper.rb
@@ -2,41 +2,45 @@ require 'openssl'
 require 'yaml'
 
 class SecretKeeper
-  def self.hi
-    puts '...(said nothing)'
-  end
-
   def self.encrypt_files
     puts 'Encrypting...'
+    ok_queue = []
     sk = SecretKeeper.new
     sk.tasks.each do |task|
       from = task['encrypt_from']
       to = task['encrypt_to']
 
       result = sk.encrypt_file(from, to)
+      ok_queue << result if result == :ok
       puts "  * #{from} --> #{to}, #{result}"
     end
     puts 'Done!'
+    ok_queue.count == sk.tasks.count
   end
 
   def self.decrypt_files
     puts 'Decrypting...'
+    ok_queue = []
     sk = SecretKeeper.new
     sk.tasks.each do |task|
       from = task['decrypt_from'] || task['encrypt_to']
       to = task['decrypt_to'] || task['encrypt_from']
 
       result = sk.decrypt_file(from, to)
+      ok_queue << result if result == :ok
       puts "  * #{from} --> #{to}, #{result}"
     end
     puts 'Done!'
+    ok_queue.count == sk.tasks.count
   end
 
   def initialize
+    fail 'environment variable OPENSSL_PASS not exist' if ENV['OPENSSL_PASS'].nil?
     env = ENV['RAILS_ENV'] || 'development'
     string = File.open('config/secret-keeper.yml', 'rb') { |f| f.read }
     config = YAML.load(string)[env]
-    @password = ENV['OPENSSL_PASS'] || 'DEFAULT-PASSWORD'
+    fail 'config/secret-keeper.yml incorrect or environment not exist' if config.nil?
+    
     @tasks = config['tasks']
     @using_cipher = OpenSSL::Cipher.new(config['cipher'])
   end
@@ -50,7 +54,7 @@ class SecretKeeper
     File.open(to_file, 'w:ASCII-8BIT') { |f| f.write(encrypted) }
     :ok
   rescue => e
-    "fail: #{e}"
+    e
   end
 
   def decrypt_file(from_file, to_file)
@@ -58,20 +62,20 @@ class SecretKeeper
     File.open(to_file, 'w') { |f| f.write(decrypted) }
     :ok
   rescue => e
-    "fail: #{e}"
+    e
   end
 
   private
 
   def encrypt(data)
     cipher = @using_cipher.encrypt
-    cipher.key = Digest::SHA1.hexdigest(@password)
+    cipher.key = Digest::SHA1.hexdigest(ENV['OPENSSL_PASS'])
     cipher.update(data) + cipher.final
   end
 
   def decrypt(data)
     cipher = @using_cipher.decrypt
-    cipher.key = Digest::SHA1.hexdigest(@password)
+    cipher.key = Digest::SHA1.hexdigest(ENV['OPENSSL_PASS'])
     cipher.update(data) + cipher.final
   end
 end
